Add HomePage object and refactor openPage.spec.ts
Added homePO.ts and page.ts in pageobjects directory to structure the code in an organized and scalable manner. This modularizes the way we handle actions on the homepage and abstracts these actions into reusable methods. Furthermore, openPage.spec.ts has been refactored to leverage these newly created page object models, enhancing the readability and maintainability of the test script.