Rename and update import paths in pageobjects and specs
The necessity of this change arose from the an issue where the older notations for file's import path were causing errors during test execution. The ".ts" extension was updated to ".js", and files were renamed with a ".page" suffix for better clarity. Also, a redundant 'beforeEach' block in 'openPage.spec.ts' was removed to clean up the code.